STATE_ -> enum class State
The sys.state variable is always set with =, never |=,
so the values are guaranteed to be unique.  It was tested
sometimes by OR-ing with bitmasks, presumably for efficiency,
but that mixed usage led to code that could be confusing.
Splitting out those tests into chained comparisons is probably
marginally less efficient than the bitmask method, but in all
likelihood, not enough to matter.