Fixed $system/control=reset problem
It was fairly insidious - the client read loop alway thought that
serial data is available, due to a problem with uint8_t vs int.
With uint8_t, there was no way to receive a -1 return value from
the various low level read routines.

I also fixed an ordering problem with config vs client input.
It is necessary to do the config first because client input
depends on the config.