Fixed bug with null string values on config items ...
The actual fix was a simple two-line addition to Tokenizer.cpp, but
in the process, I discovered that the item(... StringRange& ) overload
made no sense since there are no StringRange-type configuration items.
StringRange is used between the Tokenizer and Parser, and not outside
that domain.  So I eliminated the StringRange item overload in favor
of directly implementing item(... String& ).