Fixed spurious speed map message
To repro the problem, use a no-spindle config then send $/board .
It will show  [MSG:ERR: Speed map max speed is 0. Using default]

The fix is twofold.  First, we change the displayer so it does
not run afterParse when a tree item is displayed but not changed.
Second, we fix NullSpindle so it has an empty speed map, instead
of the kind of map that results in the map being defaulted.