ControlPin rework
1. Control pins are now registered in a vector instead of being
named individually.  That makes it easier to operate on them as a
group.
2. Renamed report() to report_status() to reflect the use of that
routine in formatting status reports and to avoid confusion with
the other report() that does something different.
3. Renamed system_check_safety_door_ajar() to safety_door_ajar()
to avoid excessive wordiness and to fix the problem that the system_
prefix is supposed to mean that the function is defined in System.cpp
4. Added debouncing to control pins to avoid rapid-fire event
assertions and as a model for other debouncing code should we need it.