Fix #668 - erratic step pulse widths in RMT mode
The problem is that the rmt_config structure was unitialized,
leading to random values in the RMB_IDLE_OUT_LC_CH bit.  When
that bit happened to be 1, it worked correctly,  When it happened
to be 0, the pulse length was wrong.

I am not entirely sure how this plays out because that bit in
question is set explicitly in later code, but empirically, if
we initialize the structure first, the value in that bit becomes
predictable and correct, whereas before, it would change from
reboot to reboot.