Fix reset from arc
If an arc is in progress when a feedhold followed by
a ctrl-x reset comes in, the next move after the reset
is wrong.  For example:

G0 X0 Y0
G3 X0 Y0 I-100 J0 F500
!
CTRL-X
$J=G91 Z0.1 F100

Instead of the jog going only in the Z direction, there
will be motion in X and Y too.  The problem is caused
by incorrect synchronization of the GCode position after
the aborted arc.  The code incorrectly sets the GCode position
to the target of the arc, as if the arc had completed,
instead of setting it to the actual position that existed
at the time of the hold+reset.  (It is possible that the
same problem might occur without the feedhold, only the
reset.  In any case, this patch works in both scenarios.)