Fix #1196 - abnormal operating speed (#1204)
The problem is that, if you issue a feedrate override when
there is no motion, it calls the planner recalculate code
which then incorrectly updates some planner blocks that are
not in use with bogus data.  Then when you issue a jog, it
uses the bogus data.

I think this is the correct fix but it hasn't been tested with
a lot of different motion situations.  It certainly fixes the
problem as stated, and seems to work for simple non-jog G1
commands.