Separate types for axis_t, motor_t, objnum_t
Also upgrade to 9 axes based on PR #1552

By and large, uint8_t is now used primarily for bytes in a buffer,
instead of as a proxy for a number that has more semantic content.
The prototypical loop over axes is now consistently expressed as
   for (axis_t axis = X_AXIS; axis < n_axis; axis++)
It used to be a mishmash of int, size_t, and uint8_types with
various names for the index variable.  axis_t is now an enum,
which helps to find misuse problems.